requirements.txt
Flask==2.3.2
pymysql==1.1.0
python-dotenv==1.0.0